rename lib/ -> src/
