+++ .npmignore
@@ -1,5 +1,4 @@
 node_modules/
-lib/
 spec/
 .babelrc
 .eslintrc
