+++ src/adapter.js
