+++ src/allergens.js
